v3.0.0 - Organizing Tests Solution folder
