+++ src/DbBroker.Showcase.Cli/README.md
@@ -2,11 +2,16 @@
 
 The easyest way of getting started with DBBroker is running the showcase using Docker Compose to run the supported databases.
 
-**NOTE:** You will need a Personal Access Token from <container-registry.oracle.com> to run `docker login container-registry.oracle.com` before building the Oracle database image.
+The Database vendors versions are:
+
+- Oracle 21.3.0 XE
+- SQL Server 2019 - latest
+
+Start the databases by running:
 
 ```bash
-docker-compose build
-docker-compose up -d
+docker compose build
+docker compose up -d
 ```
 
 You can also build and run a single database selectively, if you prefer.
