+++ src/DbBroker.Tests.DataModels/dbbroker.config.json
@@ -2,7 +2,7 @@
   "$schema": "https://raw.githubusercontent.com/diegosiao/DBBroker/refs/heads/v3.0.0/docs/json-schema/dbbroker.config.schema.json",
   "contexts": [
     {
-      "namespace": "EShop.DataModels.SqlServer",
+      "namespace": "DbBroker.Tests.DataModels.SqlServer",
       "provider": "SqlServer",
       "connectionString": "Server=127.0.0.1,1439;Database=DbBroker;User Id=sa;Password=DBBroker_1;",
       "defaultSqlInsertTemplateTypeFullName": "DbBroker.Model.Providers.SqlServerExplicitKeySqlInsertTemplate",
@@ -16,7 +16,7 @@
       "views": []
     },
     {
-      "namespace": "EShop.DataModels.Oracle",
+      "namespace": "DbBroker.Tests.DataModels.Oracle",
       "provider": "Oracle",
       "connectionString": "user id=dbbroker;password=DBBroker_1;data source=//localhost:1529/xe;",
       "defaultSqlInsertTemplateTypeFullName": "DbBroker.Model.Providers.Oracle.OracleExplicitKeySqlInsertTemplate",
@@ -38,23 +38,7 @@
           }
         }
       ],
-      "views": [
-        {
-          "name": "UVW_ORDER_SUMMARY",
-          "splitsOn": [
-            { 
-              "column": "CUSTOMER_ID",
-              "type": "UvwOrderSummaryCustomer",
-              "collection": false
-            },
-            {
-              "column": "PRODUCT_ID",
-              "type": "UvwOrderSummaryProduct",
-              "collection": true
-            }
-          ]
-        }
-      ]
+      "views": []
     }
   ]
 }
