+++ src/DbBroker.Tests/Providers/Oracle/UpdateTests.cs
@@ -1,6 +1,6 @@
 using System.Data.Common;
 using DbBroker.Model;
-using EShop.DataModels.Oracle;
+using DbBroker.Tests.DataModels.Oracle;
 using Microsoft.Extensions.DependencyInjection;
 using Oracle.ManagedDataAccess.Client;
