+++ src/DbBroker.Tests/appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "SqlServer": "Server=localhost,1439;Database=DbBroker;User Id=sa;Password=DBBroker_1;",
+    "SqlServer": "Server=localhost,1439;Database=DbBroker;User Id=sa;Password=DBBroker_1;Encrypt=false;",
     "Oracle": "user id=dbbroker;password=DBBroker_1;data source=//localhost:1529/xe;"
   }
 }
