+++ src/DbBroker.sln
@@ -1,17 +1,24 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.5.002.0
+VisualStudioVersion = 17.5.2.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DbBroker.Cli", "DbBroker.Cli\DbBroker.Cli.csproj", "{638007D9-F403-4D4E-BB91-D366B9F4C265}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DbBroker", "DBBroker\DbBroker.csproj", "{7CFB8562-9B08-44F6-AC7A-134DABF87741}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DbBroker", "DbBroker\DbBroker.csproj", "{7CFB8562-9B08-44F6-AC7A-134DABF87741}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DbBroker.Common", "DbBroker.Common\DbBroker.Common.csproj", "{2AB2592F-0851-46CD-8B2B-25D6EA9E6C1A}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DbBroker.Showcase.Cli", "DbBroker.Showcase.Cli\DbBroker.Showcase.Cli.csproj", "{C9575870-7543-4BC2-A3A8-2DFCA270ACA7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DbBroker.Unit.Tests", "DBBroker.Unit.Tests\DbBroker.Unit.Tests.csproj", "{476FFE7A-E11E-4515-8814-DDEFBA9361F3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DbBroker.Tests.DataModels", "DbBroker.Tests.DataModels\DbBroker.Tests.DataModels.csproj", "{A67661A7-B2C5-47D3-ACEE-C0082040B025}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DbBroker.Tests", "DbBroker.Tests\DbBroker.Tests.csproj", "{1D2D9F0F-0F39-4461-8E1B-B4ECABB0E3BE}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{0AB3BF05-4346-4AA6-1389-037BE0695223}"
+	ProjectSection(SolutionItems) = preProject
+		README.md = README.md
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -35,13 +42,22 @@ Global
 		{C9575870-7543-4BC2-A3A8-2DFCA270ACA7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C9575870-7543-4BC2-A3A8-2DFCA270ACA7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C9575870-7543-4BC2-A3A8-2DFCA270ACA7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{476FFE7A-E11E-4515-8814-DDEFBA9361F3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{476FFE7A-E11E-4515-8814-DDEFBA9361F3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{476FFE7A-E11E-4515-8814-DDEFBA9361F3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{476FFE7A-E11E-4515-8814-DDEFBA9361F3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A67661A7-B2C5-47D3-ACEE-C0082040B025}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A67661A7-B2C5-47D3-ACEE-C0082040B025}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A67661A7-B2C5-47D3-ACEE-C0082040B025}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A67661A7-B2C5-47D3-ACEE-C0082040B025}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1D2D9F0F-0F39-4461-8E1B-B4ECABB0E3BE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1D2D9F0F-0F39-4461-8E1B-B4ECABB0E3BE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1D2D9F0F-0F39-4461-8E1B-B4ECABB0E3BE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1D2D9F0F-0F39-4461-8E1B-B4ECABB0E3BE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{C9575870-7543-4BC2-A3A8-2DFCA270ACA7} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
+		{A67661A7-B2C5-47D3-ACEE-C0082040B025} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
+		{1D2D9F0F-0F39-4461-8E1B-B4ECABB0E3BE} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {920DA4FA-1F6D-4774-AD4D-520C4A287E56}
