+++ src/DbBroker/Compatibility/DescriptorComparer.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace DbBroker.Compatibility;
+
+public class DescriptorComparer
+{
+
+}
