+++ src/DbBroker/Compatibility/DescriptorGenerator.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace DbBroker.Compatibility;
+
+/// <summary>
+/// This class generates database descriptors and compatibility reports
+/// </summary>
+public class DescriptorGenerator
+{
+
+}
